requirements.txt
asgiref==3.2.10
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.4
channels==3.0.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-crispy-forms==1.8.0
django-filter==2.3.0
django-filters==0.2.1
django-formtools==2.1
django-guardian==2.3.0
django-import-export==2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.9.2
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.7.0
incremental==17.5.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.2.2
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==20.0.1
pytz==2019.3
PyYAML==5.2
requests==2.22.0
service-identity==18.1.0
six==1.15.0
slugify==0.0.1
sqlparse==0.3.1
tablib==0.14.0
Twisted==20.3.0
txaio==20.12.1
uritemplate==3.0.1
urllib3==1.25.7
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.2.0
